Convert milliseconds to HCI units without an undefined narrowing
mSecTo0_625 cast the quotient to uint16_t. Any interval from 40960 ms up
produces a value uint16_t cannot represent, and a float to integer conversion
that does not fit is undefined. It showed: the same input gave 65535 folded at
compile time and 0 at run time, in one binary. 41000 ms came out as 64 units,
a 40 ms interval, and BtAppAdvInit returned success. mSecTo10Ms truncated the
same way, turning 660000 ms into 464 units.

40960 ms is not an unreasonable request. Core Vol 4 Part E 7.8.53 gives the
primary advertising interval a range of 0x000020 to 0xFFFFFF, 20 ms to
10485.759375 s, so the whole range above 10.24 s was unreachable and the guard
meant to police it could never fire.

Only the destination width changes. The divide stays single precision: 0.625
is exact in binary and the widest of these fields holds 0xFFFFFF units, under
2 to the 24, so every interval in range converts exactly and 7.5 ms still
gives 12. On a Cortex-M4F that is one VDIV.F32 and one VCVT with two VCMPE.F32
range guards, all hardware, no calls and no stack frame.

Each caller applies the range its own command defines:

- Advertising interval, bt_adv_hci.cpp. 7.8.5 allows 0x0020 to 0x4000 for a
  legacy set and 7.8.53 allows 0x000020 to 0xFFFFFF for an extended one. Both
  bounds already existed; the upper one can now be reached.
- Advertising duration, bt_adv_hci.cpp. 7.8.56 gives 0x0000 the meaning
  "advertise until the Host disables the set", so a nonzero request that
  rounds to zero becomes one unit rather than an endless advertisement, and a
  request past the field is refused instead of wrapped.
- Scan interval and window, bt_gap_hci.cpp. 7.8.10 allows 0x0004 to 0x4000.
  BtGapClampScanParams takes the converted values in 32 bits so an over-range
  request clamps down to 0x4000. It ran after the wrap before, so 40960 ms
  became 0 and then clamped up to 0x0004: a request for a 41 second duty cycle
  turned into continuous scanning at full power.

The test pins 7.5, 11.25 and 152.5 next to the wide range cases, so neither
property can be traded for the other.

convutil.h joins the host test header list. It was absent, so editing it
rebuilt nothing and the first run of the new cases passed against the old
conversion still linked in.

+++ src/bluetooth/bt_gap_hci.cpp
@@ -148,7 +148,15 @@ static const BtHciCapabilities_t *BtGapCapabilitiesGet(BtHciDevice_t *pDev)
 // 0x0004..0x4000 (Core Vol 4 Part E 7.8.64 scan, 7.8.66 create connection) and
 // keep the window no larger than the interval. A small or misconfigured value
 // would otherwise reach the controller as Invalid HCI Command Parameters.
-static inline void BtGapClampScanParams(uint16_t *pInterval, uint16_t *pWindow)
+// Core Vol 4 Part E 7.8.10 allows 0x0004 to 0x4000 for both the scan interval
+// and the scan window, 2.5 ms to 10.24 s, and requires the window to be no
+// larger than the interval.
+//
+// Takes the converted values in 32 bits so an over-range request clamps down
+// to 0x4000. While the conversion truncated first, 40960 ms wrapped to 0 and
+// then clamped up to 0x0004, so a request for a 41 second duty cycle became
+// continuous scanning at full power, the opposite of what was asked.
+static inline void BtGapClampScanParams(uint32_t *pInterval, uint32_t *pWindow)
 {
 	if (*pInterval < 0x0004)
 	{
@@ -314,8 +322,8 @@ bool BtGapScanInit(BtGapScanCfg_t * const pCfg)
 		return false;
 	}
 
-	uint16_t scanInterval = mSecTo0_625(pCfg->Param.Interval);
-	uint16_t scanWindow   = mSecTo0_625(pCfg->Param.Duration);
+	uint32_t scanInterval = mSecTo0_625(pCfg->Param.Interval);
+	uint32_t scanWindow   = mSecTo0_625(pCfg->Param.Duration);
 	BtGapClampScanParams(&scanInterval, &scanWindow);
 	uint8_t scanType = (pCfg->Type == BTSCAN_TYPE_ACTIVE) ? 1 : 0;
 	uint8_t res;
@@ -486,8 +494,8 @@ bool BtGapConnect(BtGapPeerAddr_t * const pPeerAddr, BtGapConnParams_t * const p
 		return false;
 	}
 
-	uint16_t scanInterval = mSecTo0_625(s_ScanParams.Interval);
-	uint16_t scanWindow   = mSecTo0_625(s_ScanParams.Duration);
+	uint32_t scanInterval = mSecTo0_625(s_ScanParams.Interval);
+	uint32_t scanWindow   = mSecTo0_625(s_ScanParams.Duration);
 	BtGapClampScanParams(&scanInterval, &scanWindow);
 
 	uint8_t ownAddrType = BtGapResolveOwnAddr(pDev, pCapabilities);
